pyproject.toml
[project]
name = "flaskr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bootstrap-flask>=2.5.0",
    "coverage>=7.13.4",
    "cssselect>=1.4.0",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["flaskr"]
